+++ R/models-resnet.R
@@ -249,10 +249,11 @@ resnet <- torch::nn_module(
   model
 }
 
-#' ResNet-18 Model Architecture
+#' ResNet implementation
 #'
-#' ResNet-18 model architecture from
-#'   [Deep Residual Learning for Image Recognition](https://arxiv.org/pdf/1512.03385.pdf)
+#' ResNet models implementation from
+#'   [Deep Residual Learning for Image Recognition](https://arxiv.org/pdf/1512.03385.pdf) and later
+#'   related papers (see Functions)
 #'
 #' @param pretrained (bool): If TRUE, returns a model pre-trained on ImageNet.
 #' @param progress (bool): If TRUE, displays a progress bar of the download to
@@ -260,10 +261,64 @@ resnet <- torch::nn_module(
 #' @param ... Other parameters passed to the resnet model.
 #'
 #' @family models
-#'
+#' @name model_resnet
+#' @rdname model_resnet
+NULL
+
+#' @describeIn model_resnet ResNet 18-layer model
 #' @export
 model_resnet18 <- function(pretrained = FALSE, progress = TRUE, ...) {
   .resnet("resnet18", basic_block, c(2,2,2,2), pretrained, progress, ...)
 }
 
+#' @describeIn model_resnet ResNet 34-layer model
+#' @export
+model_resnet34 <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("resnet34", basic_block, c(3,4,6,3), pretrained, progress, ...)
+}
 
+#' @describeIn model_resnet ResNet 50-layer model
+#' @export
+model_resnet50 <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("resnet50", bottleneck, c(3,4,6,3), pretrained, progress, ...)
+}
+
+#' @describeIn model_resnet ResNet 101-layer model
+#' @export
+model_resnet101 <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("resnet101", bottleneck, c(3,4,23,3), pretrained, progress, ...)
+}
+
+#' @describeIn model_resnet ResNet 152-layer model
+#' @export
+model_resnet152 <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("resnet152", bottleneck, c(3,8,36,3), pretrained, progress, ...)
+}
+
+#' @describeIn model_resnet ResNeXt-50 32x4d model from ["Aggregated Residual Transformation for Deep Neural Networks"](https://arxiv.org/pdf/1611.05431.pdf)
+#' with 32 groups having each a width of 4.
+#' @export
+model_resnext50_32x4d <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("resnext50_32x4d", bottleneck, c(3,4,6,3), pretrained, progress, groups=32, width_per_group=4,...)
+}
+
+#' @describeIn model_resnet ResNeXt-101 32x8d model from ["Aggregated Residual Transformation for Deep Neural Networks"](https://arxiv.org/pdf/1611.05431.pdf)
+#' with 32 groups having each a width of 8.
+#' @export
+model_resnext101_32x8d <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("resnext101_32x8d", bottleneck, c(3,4,23,3), pretrained, progress, groups=32, width_per_group=8,...)
+}
+
+#' @describeIn model_resnet Wide ResNet-50-2 model from ["Wide Residual Networks"](https://arxiv.org/pdf/1605.07146.pdf)
+#' with width per group of 128.
+#' @export
+model_wide_resnet50_2 <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("wide_resnet50_2", bottleneck, c(3,4,6,3), pretrained, progress, width_per_group=64*2,...)
+}
+
+#' @describeIn model_resnet Wide ResNet-101-2 model from ["Wide Residual Networks"](https://arxiv.org/pdf/1605.07146.pdf)
+#' with width per group of 128.
+#' @export
+model_wide_resnet101_2 <- function(pretrained = FALSE, progress = TRUE, ...) {
+  .resnet("wide_resnet101_2", bottleneck, c(3,4,23,3), pretrained, progress, width_per_group=64*2,...)
+}